refactor(sight): split genai.rs into 5 focused submodules
